environment setup

+++ mmseg/datasets/lane_datasets/openlane.py
@@ -113,7 +113,7 @@ class OpenlaneDataset(Dataset):
         x_max = self.top_view_region[1, 0]  # 10
         self.x_min = x_min  # -10
         self.x_max = x_max  # 10
-        self.anchor_y_steps = np.array(y_steps, dtype=np.float)
+        self.anchor_y_steps = np.array(y_steps, dtype=np.float32)
         self.anchor_len = len(self.anchor_y_steps)
 
         self.y_min = self.top_view_region[2, 1]
